+++ src/app/compare/CompareSuggestions.tsx
@@ -44,8 +44,8 @@ export default memo(function CompareSuggestions({
     if (!nextCompareButton) {
       return !keptPenultimateButton;
     }
-    // skip empty buttons
-    if (!compareButton.items.length) {
+    // skip empty buttons or buttons that only contain the example item (except the first item-specific button)
+    if (compareButton.items.length < 2 && !compareButton.query.includes('name:')) {
       return false;
     }
     // if the next button has [all of, & only] the exact same items in it